fix method for switch driver and shift status

+++ src/main/java/org/example/logisticapplication/domain/Driver/Driver.java
@@ -14,7 +14,7 @@ public record Driver(
         Long  personNumber,
 
         @JsonProperty("numOfHoursWorked")
-        Integer numberOfHoursWorked,
+        Double numberOfHoursWorked,
 
         String status,
